requirements.txt
Flask
PyMuPDF
python-docx
Werkzeug
python-dateutil
dateparser
reportlab